Rename convert to common and improve license header descriptions
