+++ src/routes/settings/common.rs
@@ -2,7 +2,8 @@
 *  (c) 2018 Pi-hole, LLC (https://pi-hole.net)
 *  Network-wide ad blocking via your own hardware.
 *
-*  Convert booleans returned as strings.
+*  API
+*  Common functions for settings endpoints
 *
 *  This file is copyright under the latest version of the EUPL.
 *  Please see LICENSE file for your rights under this license. */
