+++ src/routes/settings/mod.rs
@@ -1,11 +1,21 @@
+/* Pi-hole: A black hole for Internet advertisements
+*  (c) 2018 Pi-hole, LLC (https://pi-hole.net)
+*  Network-wide ad blocking via your own hardware.
+*
+*  API
+*  Setting endpoints
+*
+*  This file is copyright under the latest version of the EUPL.
+*  Please see LICENSE file for your rights under this license. */
+
 mod get_dhcp;
 mod get_dns;
 mod get_ftldb;
 mod get_network;
-mod convert;
+mod common;
 
 pub use self::get_network::*;
 pub use self::get_ftldb::*;
 pub use self::get_dns::*;
 pub use self::get_dhcp::*;
-pub use self::convert::*;
+pub use self::common::*;
